src/main/webapp/app/entities/type-sol/route/type-sol-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ITypeSol, TypeSol } from '../type-sol.model';
import { TypeSolService } from '../service/type-sol.service';

import { TypeSolRoutingResolveService } from './type-sol-routing-resolve.service';

describe('TypeSol routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TypeSolRoutingResolveService;
  let service: TypeSolService;
  let resultTypeSol: ITypeSol | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(TypeSolRoutingResolveService);
    service = TestBed.inject(TypeSolService);
    resultTypeSol = undefined;
  });

  describe('resolve', () => {
    it('should return ITypeSol returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTypeSol = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTypeSol).toEqual({ id: 123 });
    });

    it('should return new ITypeSol if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTypeSol = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTypeSol).toEqual(new TypeSol());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as TypeSol })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTypeSol = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTypeSol).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
